backend/requirements.txt
Django==2.2.19
django-colorfield==0.7.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djoser==2.0.5
drf-extra-fields==3.4.0
flake8==4.0.1
gunicorn==20.0.4
isort==5.10.1
mccabe==0.6.1
Pillow==9.2.0
psycopg2-binary==2.8.6
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
